requirements.txt
streamlit==1.38.0
ultralytics==8.2.83
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.25.0
torch==2.4.0